--- controllers/user.js
@@ -1,49 +1,0 @@
-const sequelize = require("../config/connection")
-const User = require("../model/User");
-
-const seedDatabase = async() => {
-    await sequelize.sync();
-}
-
-exports.user_login = async function (req,res)
-{
-    const findUser = await User.findOne(
-        {
-            where: 
-            {
-            email: req.body.email,
-            //password: req.body.email
-            }
-        });
-
-    if(findUser === null)
-    {
-        console.log("invalid credentials");
-    }
-    else
-    {
-        res.send("OK");
-    }
-}
-
-exports.user_signup = async function (req,res)
-{
-    const arr = await User.findOrCreate(
-    {
-        where: 
-        {
-        username: req.body.username, 
-        email: req.body.email,
-        password: req.body.password
-        }
-    })
-
-    const wasCreated = arr[1] // the second element tells us if the instance was newly created
-
-    if(wasCreated)
-    {
-        res.send("Signup Success");
-    }
-}
-
-seedDatabase();
