+++ models/index.js
@@ -1,5 +1,6 @@
 const favouriteRecipies = require("./favouriteRecipies");
 const userProfile = require("./userProfile");
+const User = require("./User");
 
 // define relationships between the tables 
 
@@ -7,4 +8,5 @@ module.exports =
 {
     favouriteRecipies, 
     userProfile,
+    User,
 };
